refactor(font): remove the PDFBox-typed API and FontShowcase from the core font package
Ahead of the 2.0 module split, take the core font package a step closer to
backend-neutral.

- DefaultFonts no longer declares library(PDDocument) or
  library(PDDocument, Collection<FontFamilyDefinition>). They were thin
  delegators to document.backend.fixed.pdf.PdfFontLibraryFactory, which already
  exposes the identical methods; the (test-scope) callers now call the factory
  directly. compose.font is now free of org.apache.pdfbox.
- FontShowcase moves from compose.font to document.backend.fixed.pdf: it
  composes a preview through the document DSL and renders it, so it belongs with
  the PDF backend it rides. Its two tests move alongside, and its public methods
  gain javadoc (the pdf package is javadoc-gated with doclint=all; the font
  package was not). GraphCompose.renderAvailableFontsPreview() imports it from
  the new home.

DefaultFonts.standardLibrary() and its EntityManager caller are intentionally
left untouched; that remaining font-to-backend delegation is cut when the
render/measurement seam lands. Behaviour is unchanged: the FontShowcase
render/snapshot tests and every layout snapshot pass with zero baseline updates.
