+++ src/test/java/com/demcha/compose/document/backend/fixed/pdf/FontShowcaseRenderTest.java
@@ -1,6 +1,7 @@
-package com.demcha.compose.font;
+package com.demcha.compose.document.backend.fixed.pdf;
 
 import com.demcha.compose.GraphCompose;
+import com.demcha.compose.font.FontName;
 import com.demcha.testing.VisualTestOutputs;
 import org.apache.pdfbox.Loader;
 import org.apache.pdfbox.pdmodel.PDDocument;
